+++ pypgx/__init__.py
@@ -54,6 +54,7 @@ from .api.utils import (
     predict_cnv,
     prepare_depth_of_coverage,
     print_metadata,
+    slice_bam,
     test_cnv_caller,
     train_cnv_caller,
 )