+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog
 -----------------------
 
 * Add new optional arguments ``--genes`` and ``--exclude`` to :command:`prepare-depth-of-coverage` command.
+* Add new command :command:`slice-bam`.
 
 0.14.0 (2022-04-03)
 -------------------